+++ internal/usecase/user_test.go
@@ -2,59 +2,10 @@ package usecase
 
 import (
 	"github.com/Ptt-official-app/Ptt-backend/internal/config"
-	"github.com/Ptt-official-app/Ptt-backend/internal/repository"
-
-	"github.com/PichuChen/go-bbs"
 
 	"context"
 	"testing"
-	"time"
 )
-
-// implements repository.Repository
-type MockRepository struct{}
-
-// repository/board.go
-func (repo *MockRepository) GetBoards(ctx context.Context) []bbs.BoardRecord {
-	return []bbs.BoardRecord{}
-}
-func (repo *MockRepository) GetBoardArticle(ctx context.Context, boardID, filename string) ([]byte, error) {
-	return []byte{}, nil
-}
-func (repo *MockRepository) GetBoardArticleRecords(ctx context.Context, boardID string) ([]bbs.ArticleRecord, error) {
-	return []bbs.ArticleRecord{}, nil
-}
-
-func (repo *MockRepository) GetBoardTreasureRecords(ctx context.Context, boardID string, treasureIDs []string) ([]bbs.ArticleRecord, error) {
-	return []bbs.ArticleRecord{}, nil
-}
-
-// repository/user.go
-func (repo *MockRepository) GetUsers(ctx context.Context) []bbs.UserRecord {
-	ret := []bbs.UserRecord{}
-	ret = append(ret, &MockUser{
-		userId: "pichu",
-	})
-	return ret
-}
-func (repo *MockRepository) GetUserFavoriteRecords(ctx context.Context, userID string) ([]bbs.FavoriteRecord, error) {
-	return []bbs.FavoriteRecord{}, nil
-}
-
-type MockUser struct {
-	userId string
-}
-
-func (u *MockUser) UserId() string                       { return u.userId }
-func (u *MockUser) HashedPassword() string               { return "" }
-func (u *MockUser) VerifyPassword(password string) error { return nil }
-func (u *MockUser) Nickname() string                     { return "" }
-func (u *MockUser) RealName() string                     { return "" }
-func (u *MockUser) NumLoginDays() int                    { return 0 }
-func (u *MockUser) NumPosts() int                        { return 0 }
-func (u *MockUser) Money() int                           { return 0 }
-func (u *MockUser) LastLogin() time.Time                 { return time.Unix(0, 0) }
-func (u *MockUser) LastHost() string                     { return "" }
 
 func TestGetUserByID(t *testing.T) {
 
@@ -85,6 +36,3 @@ func TestGetUserByID(t *testing.T) {
 	}
 
 }
-
-// To make sure MockRepository implement repository.Repository
-var _ repository.Repository = &MockRepository{}
